apps/ingestion-worker/requirements.txt
pynsq==0.9.1
crawl4ai
docling
python-snappy
uvloop
pydantic
pydantic-settings
pytest
pytest-asyncio
tornado